+++ src/main/java/com/appcenter/BJJ/domain/item/dto/DetailItemRes.java
@@ -17,7 +17,7 @@ public class DetailItemRes extends ItemRes {
 
     private Boolean isOwned;
 
-    public DetailItemRes(Integer itemId, String itemName, ItemType itemType, ItemLevel itemLevel, String imageName,
+    public DetailItemRes(Long itemId, String itemName, ItemType itemType, ItemLevel itemLevel, String imageName,
                          LocalDateTime validPeriod, boolean isWearing, boolean isOwned) {
         super(itemId, itemName, itemType, itemLevel, imageName);
         this.validPeriod = validPeriod;
